完成pipeline整个通道的修改和调试工作，可以完成长时间映射的保持，同时解决了之前传输跑着跑着就卡着了的bug，解决了crash的bug；完成了图形指令的编解码的调试，图形指令的翻译仍未测试

+++ include/direct-express/express_device_common.h
@@ -3,8 +3,8 @@
 
 #include "direct-express/direct_express_distribute.h"
 
-#define EXPRESS_LOG_FUN_ID 0
-#define EXPRESS_GPU_FUN_ID 1
+#define EXPRESS_LOG_FUN_ID ((unsigned long long)0)
+#define EXPRESS_GPU_FUN_ID ((unsigned long long)1)
